Move I2C driver to i2c::master
