+++ esp-hal/src/i2c/mod.rs
@@ -1,0 +1,11 @@
+//! # Inter-Integrated Circuit (I2C)
+//!
+//! I2C is a serial, synchronous, multi-device, half-duplex communication
+//! protocol that allows co-existence of multiple masters and slaves on the
+//! same bus. I2C uses two bidirectional open-drain lines: serial data line
+//! (SDA) and serial clock line (SCL), pulled up by resistors.
+
+pub mod master;
+
+#[cfg(lp_i2c0)]
+pub mod lp_i2c;
